fixing value type to double(2)

+++ gateway/src/main/java/com/hurence/webapiservice/historian/util/models/ResponseAsList.java
@@ -36,7 +36,7 @@ public class ResponseAsList {
         public SubResponse(JsonArray line) {
             this.metric = line.getString(0);
             this.date = line.getLong(2);
-            this.value = line.getLong(1);
+            this.value = line.getDouble(1);
         }
     }
